requirements.txt
streamlit
pandas
numpy
plotly
python-dotenv
google-genai
growwapi
groq
requests